Добавлено несколько городов Добавлена поддержка DiffUtils Остался фрагмент и комментарии

+++ app/src/main/java/home/stanislavpoliakov/meet13_practice/Weather.java
@@ -29,5 +29,5 @@ public class Weather {
 
     @Embedded (prefix = "flags")
     public WFlags flags;
-    public int offset;
+    public double offset;
 }
